--- a/ui/templates/main.html
+++ b/ui/templates/main.html
@@ -43,7 +43,7 @@
     or run until interrupted:<input type="checkbox" name="t" onchange="toggleDuration(this)" />
     or run for exactly <input type="text" name="n" size="6" value="" /> calls. <br />
     Threads/Simultaneous connections: <input type="text" name="c" size="6" value="8" /> <br />
-    Request jitter:<input type="checkbox" name="reqjitter" /> <br />
+    Jitter:<input type="checkbox" name="jitter" /> <br />
     Percentiles: <input type="text" name="p" size="20" value="50, 75, 90, 99, 99.9" /> <br />
     Histogram Resolution: <input type="text" name="r" size="8" value="0.0001" /> <br />
     Headers: <br />