We have been using `fortio` to measure Istio's latency. One of the problems with its long-tail latency issue is the benchmark itself. When we specify `-c 100`, 100 goroutines are created to send off 100 requests all at the same time (using tcpdump, we saw these requests all got to the server side within 100us). So on the server side, it sees these mini-bursts of requests over and over again. Due to Envoy worker threads being single threaded, this creates a significant queuing delay. In the real world, bursts of requests do happen, but we should not artificially create thundering herd problem to make Istio (or whatever is being tested) to have latency problem that is worse than what it really is. 

A flag `--request-jitter` is added to add (+/-)10% jitter between requests of parallel clients. Here is the result measuring Istio latency with and without the flag:

* Without added jitter:
```
# ./fortio load -qps 1000 -c 100 -t 300s 10.97.103.115:8079
Fortio 1.3.2-pre running at 1000 queries per second, 4->4 procs, for 5m0s: 10.97.103.115:8079
16:33:26 I httprunner.go:82> Starting http test for 10.97.103.115:8079 with 100 threads at 1000.0 qps
16:33:26 W http_client.go:142> Assuming http:// on missing scheme for '10.97.103.115:8079'
Starting at 1000 qps with 100 thread(s) [gomax 4] for 5m0s : 3000 calls each (total 300000)
Ended after 5m0.013834356s : 300000 calls. qps=999.95
Sleep times : count 299900 avg 0.088003375 +/- 0.00315 min 0.059958014 max 0.098711726 sum 26392.2122
Aggregated Function Time : count 300000 avg 0.011201918 +/- 0.003032 min 0.001004862 max 0.038903617 sum 3360.57543
# range, mid point, percentile, count
>= 0.00100486 <= 0.002 , 0.00150243 , 0.01, 19
> 0.002 <= 0.003 , 0.0025 , 0.02, 33
> 0.003 <= 0.004 , 0.0035 , 0.04, 75
> 0.004 <= 0.005 , 0.0045 , 0.07, 96
> 0.005 <= 0.006 , 0.0055 , 0.27, 601
> 0.006 <= 0.007 , 0.0065 , 3.53, 9766
> 0.007 <= 0.008 , 0.0075 , 14.33, 32409
> 0.008 <= 0.009 , 0.0085 , 28.50, 42496
> 0.009 <= 0.01 , 0.0095 , 41.03, 37599
> 0.01 <= 0.011 , 0.0105 , 52.01, 32940
> 0.011 <= 0.012 , 0.0115 , 62.61, 31792
> 0.012 <= 0.014 , 0.013 , 81.53, 56771
> 0.014 <= 0.016 , 0.015 , 93.95, 37256
> 0.016 <= 0.018 , 0.017 , 97.99, 12126
> 0.018 <= 0.02 , 0.019 , 99.12, 3370
> 0.02 <= 0.025 , 0.0225 , 99.90, 2349
> 0.025 <= 0.03 , 0.0275 , 100.00, 300
> 0.03 <= 0.035 , 0.0325 , 100.00, 1
> 0.035 <= 0.0389036 , 0.0369518 , 100.00, 1
# target 50% 0.0108168
# target 75% 0.0133096
# target 90% 0.0153637
# target 99% 0.0197929
# target 99.9% 0.0250333
Sockets used: 100 (for perfect keepalive, would be 100)
Code 200 : 300000 (100.0 %)
Response Header Sizes : count 300000 avg 295.00439 +/- 0.06609 min 295 max 296 sum 88501316
Response Body/Total Sizes : count 300000 avg 295.00439 +/- 0.06609 min 295 max 296 sum 88501316
All done 300000 calls (plus 100 warmup) 11.202 ms avg, 1000.0 qps
```

* With the `--request-jitter` flag:
```
# ./fortio load -qps 1000 -c 100 -t 300s --request-jitter 10.97.103.115:8079
Fortio 1.3.2-pre running at 1000 queries per second, 4->4 procs, for 5m0s: 10.97.103.115:8079
16:28:15 I httprunner.go:82> Starting http test for 10.97.103.115:8079 with 100 threads at 1000.0 qps
16:28:15 W http_client.go:142> Assuming http:// on missing scheme for '10.97.103.115:8079'
Starting at 1000 qps with 100 thread(s) [gomax 4] for 5m0s : 3000 calls each (total 300000)
Ended after 5m0.010714621s : 300000 calls. qps=999.96
Sleep times : count 299900 avg 0.097020439 +/- 0.00854 min 0.054097274 max 0.119709755 sum 29096.4297
Aggregated Function Time : count 300000 avg 0.0027441949 +/- 0.002333 min 0.000662307 max 0.031994066 sum 823.258457
# range, mid point, percentile, count
>= 0.000662307 <= 0.001 , 0.000831154 , 7.27, 21823
> 0.001 <= 0.002 , 0.0015 , 54.81, 142595
> 0.002 <= 0.003 , 0.0025 , 73.09, 54853
> 0.003 <= 0.004 , 0.0035 , 80.92, 23490
> 0.004 <= 0.005 , 0.0045 , 86.02, 15301
> 0.005 <= 0.006 , 0.0055 , 90.61, 13770
> 0.006 <= 0.007 , 0.0065 , 93.94, 9987
> 0.007 <= 0.008 , 0.0075 , 96.14, 6611
> 0.008 <= 0.009 , 0.0085 , 97.51, 4090
> 0.009 <= 0.01 , 0.0095 , 98.32, 2426
> 0.01 <= 0.011 , 0.0105 , 98.80, 1450
> 0.011 <= 0.012 , 0.0115 , 99.12, 960
> 0.012 <= 0.014 , 0.013 , 99.55, 1281
> 0.014 <= 0.016 , 0.015 , 99.77, 676
> 0.016 <= 0.018 , 0.017 , 99.86, 266
> 0.018 <= 0.02 , 0.019 , 99.92, 168
> 0.02 <= 0.025 , 0.0225 , 99.98, 179
> 0.025 <= 0.03 , 0.0275 , 100.00, 62
> 0.03 <= 0.0319941 , 0.030997 , 100.00, 12
# target 50% 0.00189889
# target 75% 0.00324389
# target 90% 0.00586696
# target 99% 0.0116292
# target 99.9% 0.0194405
Sockets used: 100 (for perfect keepalive, would be 100)
Code 200 : 300000 (100.0 %)
Response Header Sizes : count 300000 avg 295.00006 +/- 0.007528 min 295 max 296 sum 88500017
Response Body/Total Sizes : count 300000 avg 295.00006 +/- 0.007528 min 295 max 296 sum 88500017
All done 300000 calls (plus 100 warmup) 2.744 ms avg, 1000.0 qps
```

The latency difference is pretty significant with or without jitter under the same test condition (1000 qps, 100 clients, 300 seconds). Avg latency went from 11.2ms to 2.7ms. Long-tail latency also improved significantly. This problem with request jitter has been discussed here as well: https://github.com/envoyproxy/envoy/issues/5536