@ldemailly 2019-04-26
Thanks for the PR ! I would like to question a bit the premise though:

If you don't want 100 near simultaneous incoming connections, why use `-c 100` ? the idea of `-c` is indeed to have that many connections, also the initial request isn't counted in the stats and after the initial request there is already some jitter because the network interface on the client can only do 1 thing at a time

Unless you have very large latencies, you don't need 100 connections to get 1000 qps ? is the above reproducible multiple times ? what happens with -c 10 ?