@ldemailly 2019-04-26
Can you show me the sleep times (-loglevel verbose) with and without jitter ?
Also add the flag to the report so you can know if jitter was on or off when looking at/graphing the reports
And can you add it to the UI as well ? (checkbox)