@huang195 2019-04-26
@ldemailly thanks for the quick reply. 

> If you don't want 100 near simultaneous incoming connections, why use -c 100 ?

Because each network connection could impose some per-connection overhead. This will depend on the backend and/or proxy implementation, e.g., multi-process vs. multi-threading vs. single-threaded running an event loop. The default value of the new flag is false, which allows one to generate a load with that many simultaneous requests. However, as mentioned above, this might not reflect real world request pattern, where all clients are colluding to send a request all at exactly the same time. 

> after the initial request there is already some jitter because the network interface on the client can only do 1 thing at a time

Well, I tried that. I added a fixed amount of delay between the start of each goroutine. I the first few rounds, I could see requests are indeed staggered. But with 100 connections, unless I have 100 dedicated CPUs, there will be drift in their execution time. What I saw was eventually, the 100 goroutine became "synchronized" again and started to send all their requests in batches as if the initial delay added was never there. This is the reason the jitter needs to be enforced each time a client goes to sleep.